Move example App to its own folder
