+++ react-app/src/components/TestApp/index.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import logo from './logo.svg';
+import styles from "./index.module.scss";
+
+class App extends Component {
+  render() {
+    return (
+      <div className={styles.container}>
+        <header className={styles.header}>
+          <img src={logo} className={styles.logo} alt="logo" />
+          <h1 className={styles.title}>Welcome to React</h1>
+        </header>
+        <p className={styles.intro}>
+          To get started, edit <code>src/App.js</code> and save to reload.
+        </p>
+      </div>
+    );
+  }
+}
+
+export default App;
