+++ react-app/src/components/TestApp/index.module.scss
@@ -1,28 +1,28 @@
-.App {
+.container {
   text-align: center;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
+.logo {
+  animation: logo-spin infinite 20s linear;
   height: 80px;
 }
 
-.App-header {
+.header {
   background-color: #222;
   height: 150px;
   padding: 20px;
   color: white;
 }
 
-.App-title {
+.title {
   font-size: 1.5em;
 }
 
-.App-intro {
+.intro {
   font-size: large;
 }
 
-@keyframes App-logo-spin {
+@keyframes logo-spin {
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
