+++ react-app/src/components/TestApp/test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
