+++ react-app/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './components/TestApp';
 import registerServiceWorker from './registerServiceWorker';
 
 const startApp = () => {
@@ -9,7 +9,7 @@ const startApp = () => {
   registerServiceWorker();
 }
 
-if( !window.cordova ) {
+if (!window.cordova) {
   startApp();
 } else {
   document.addEventListener('deviceready', startApp, false);
